\name{XVectorList-class}
\docType{class}

\alias{class:GroupedIRanges}
\alias{GroupedIRanges-class}
\alias{GroupedIRanges}

\alias{class:XVectorList}
\alias{XVectorList-class}
\alias{XVectorList}

\alias{parallel_slot_names,GroupedIRanges-method}
\alias{parallel_slot_names,XVectorList-method}

\alias{as.data.frame,GroupedIRanges-method}
\alias{show,GroupedIRanges-method}

\alias{width,XVectorList-method}
\alias{elementNROWS,XVectorList-method}
\alias{names,XVectorList-method}
\alias{names<-,XVectorList-method}
\alias{[,XVectorList-method}
\alias{subseq,XVectorList-method}
\alias{bindROWS,XVectorList-method}

\alias{showAsCell,XVectorList-method}

\alias{unsplit_list_of_XVectorList}


\title{XVectorList objects}

\description{
  THIS IS A WORK-IN-PROGRESS!!


  An XVectorList object is *conceptually* a list of \link{XVector}
  objects.
}

\author{H. Pagès}

\seealso{
  \link{XVector-class},
  \link{XRawList-class},
  \code{\link{compact}}
}

\keyword{methods}
\keyword{classes}
